+++ lua/obsidian/commands/new_from_template.lua
@@ -4,38 +4,40 @@ local log = require "obsidian.log"
 ---@param client obsidian.Client
 ---@param data CommandArgs
 return function(client, data)
-  if not client:templates_dir() then
-    log.err "Templates folder is not defined or does not exist"
-    return
-  end
-
   local picker = client:picker()
   if not picker then
     log.err "No picker configured"
     return
   end
 
-  ---@type obsidian.Note
-  local note
-  if data.args and data.args:len() > 0 then
-    note = client:create_note { title = data.args, no_write = true }
-  else
-    local title = util.input("Enter title or path (optional): ", { completion = "file" })
+  local title = data.fargs[1]
+  local template = data.fargs[2]
+
+  if title ~= nil and template ~= nil then
+    local note = client:create_note { title = title, template = template, no_write = false }
+    client:open_note(note, { sync = true })
+    return
+  end
+
+  if title == nil or title == "" then
+    title = util.input("Enter title or path (optional): ", { completion = "file" })
     if not title then
       log.warn "Aborted"
       return
     elseif title == "" then
       title = nil
     end
-    note = client:create_note { title = title, no_write = true }
   end
-
-  -- Open the note in a new buffer.
-  client:open_note(note, { sync = true })
 
   picker:find_templates {
     callback = function(name)
-      client:write_note_to_buffer(note, { template = name })
+      if name == nil or name == "" then
+        log.warn "Aborted"
+        return
+      end
+      ---@type obsidian.Note
+      local note = client:create_note { title = title, template = name, no_write = false }
+      client:open_note(note, { sync = false })
     end,
   }
 end